[refactor] SortDirection을 enums 패키지로 이동
